Some housekeeping done on the attributes of the box layout and each of the widgets to better organise them.
Icon attributes were added to the button widget in preparation for the integration of icons into the widget.

Default icons were also added to a new Icons directory in the Resources directory. The idea is that the toolkit will come with these default icons that the user can reference by name (in a string).
For additional icons the user will have the option to either add icons to that folder or to reference the icon by its path.