After many attempts I finally found a good and simple way to change the color of an icon using the pygame.BLEND tools.
While it looks quite good there are some visual inconsistencies in the icons where a light grey border seems present.
I believe this occurs when there a high alpha pixels (not very transparent) that near black but are still translucent. This can obviously be fixed by going through all of the icons and modifying them either by hand or with a program.