Fixed error where Error was improperly called
Introduced Icon fix program to Resources. This is a single time run file that was intended to reduce the alpha values of high alpha value pixels in the hope of fixing the minor visual inconsistencies with the new icons.

However, while it worked in some ways it failed in others as it sometimes thinned the icons to the point that it was better before.

I even tried doing a threshold system in which pixels with an alpha value below or above 150 would become either transparent or opaque. However, for some reason this didn't soleve the problem and also failed to maintain the icon's original quality.

I guess I'll have to live with it until I find a better way of dealing with it so that I can move on to more important ideas