Major refactoring done to the app.py file to allow for scrolling layouts as well as removing large sections of redundant code.
Every layout is given an intermediate_layout which is a surface on which the real size of the layout is blitted (real size is size * 1 by default).
The user can choose to change the real size and enable scroll.

On top of this, within the app.py file each layout is allocated a surface on the screen which works to limit the size of each layout regardless of the layout's real size. As a result of this, the layouts themselves can now all work with coordinates relative to 0 and can ignore their place in the screen.

This also applies to widgets.

Some things still haven't been completed such as activating certain layouts so that only that layout is scrolled through rather than all of them.

Additionally, not everything will work for a while as there are bound to be countless bugs as a result of this refactor or maybe I'm not actually terrible at python :).