Changed the way that the layout handles the width and height of layouts and implemented a size and position hint for the widgets so that you can resize them and position them within their own allocated areas. This clashes slightly with the align features and scale features so that'll need to be fixed. Still various bugs due to the relatively complicated refactor with the widget dimensions and positions.
low FPS becomes very apparent with the scroll feature and so the image loading is set to only be performed once to save on resources. This boosted it from 20 FPS with two images on screen to around 60.

Seeing as this is only running on 1 core I'll need to optimise further as creating a more complex GUI will have a bigger impact on FPS.

Added the showcase directory which houses Showcase1.py which shows almost all of the features from the widgets to the new scrolling features.