A lots been added since the last commit simply because the text widget stopped working because of some of optimisations.
Speaking of some optimisations, all of the widgets have been reworked so that required objects, surfaces and rendered texts are loaded beforehand. This improves performance of Showcase1.py from around 60fps to around 300fps.
Changes were made to the wrap_text.py to accomodate for this. This included replicating the code into two separate functions in which one of them (static) returns the list of words and positions so that it can be cached by the widget and blitted locally.

Additionally, unless i mentioned this in the previous commit, I added position and size hints to give the user far greater control over the alignment and position of the widgets.

I also added a new widget called the Checkbox widget which is simply a checkbox that offers the ability to add a header. This currently has a bug where the tick can disappear when scrolling.

Speaking of scrolling I enhanced it so that it moves more smoothly. However, for this to be reliable it requires a fixed frame rate but I may allow the user to unlock the frame rate in future. This works by having the layout move multiple times for every time the mouse wheel is "pressed".